requirements-stable.txt
# Stable requirements for Django project
# Recommended Python version: 3.12.x

# Core Django
Django==5.1.4
asgiref==3.8.1

# Django REST Framework
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.1
PyJWT==2.9.0

# API Documentation
drf-spectacular==0.28.0

# Database
psycopg2-binary==2.9.9

# CORS
django-cors-headers==4.6.0

# Filtering
django-filter==24.3

# Development tools
django-extensions==3.2.3

# Email verification
Django-Verify-Email==3.0.3

# Background tasks
celery==5.4.0
redis==5.2.1
kombu==5.4.2
amqp==5.3.1
billiard==4.2.1
vine==5.1.0

# Configuration
python-decouple==3.8

# Image handling
Pillow==11.0.0

# Utilities
pytz==2024.2
python-dateutil==2.9.0.post0
requests==2.32.3

# Account management
django-appconf==1.0.6

# Other dependencies
click==8.1.7
packaging==24.2
sqlparse==0.5.2
typing_extensions==4.12.2
PyYAML==6.0.2
urllib3==2.2.3
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
six==1.16.0
